% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expectvalue.R
\name{EoneD}
\alias{EoneD}
\title{Single variable Rejection Sampling}
\usage{
EoneD(f, N, lb, ub)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the nimber of attempted samples.}

\item{lb}{lower bound of support of f}

\item{ub}{upper bound of support of f}
}
\value{
mean of given pdf (with one r.v.)
}
\description{
This function 'EoneD(f,N,lb,up)' shows the Expect value of the given pdf (with one r.v.) by using single variabel rejection sampling.
}
\examples{

f<- function(x) dnorm(x,-10,2)
EoneD(f,10000, Inf, Inf)

f<- function(x) {ifelse(-1< x & x < 0, 2*(x+1), 0)}
EoneD(f,10000, -1, 0)


}
