% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{prooneD}
\alias{prooneD}
\title{Single variable Rejection Sampling}
\usage{
prooneD(f, N, lb, ub, vall = -Inf, valu = Inf)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the nimber of attempted samples.}

\item{lb}{lower bound of support of f}

\item{ub}{upper bound of support of f}

\item{vall}{lower value, default is -Inf}

\item{valu}{upper value, default is Inf}
}
\value{
probability of given pdf (with one r.v.) and situation
}
\description{
This function 'prooneD(f,N,lb,up,vall,valu)' shows the probability of the given pdf (with one r.v.) the situation could be smaller than,
large than or between any range
}
\examples{

f<- function(x) dnorm(x,-10,2)
prooneD(f,10000, Inf, Inf,vall=-10, valu=15)

f<- function(x) {ifelse(-1< x & x < 0, 2*(x+1), 0)}
prooneD(f,10000, -1, 0,valu=-0.5)


}
