% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoDsample.R
\name{twoDsample}
\alias{twoDsample}
\title{Two variable Rejection Sampling}
\usage{
twoDsample(f, N, lbx = -5000, ubx = 5000, lby = -5000, uby = 5000)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the nimber of attempted samples.}

\item{lbx}{lower bound of x support of f}

\item{ubx}{upper bound of x support of f}

\item{lby}{lower bound of y support of f}

\item{uby}{upper bound of y support of f}
}
\value{
A vector containing samples from pdf
        ggplot of simulation joint density function
}
\description{
This function "twoDsample(f, N, lbx, ubx, lby, uby)" implements two variabel rejection sampling for rvs with bounded support and which have bounded pdf.
}
\details{
Additionaly, the function 'twoDsampleplot()' shows the plot of simulation joint density function.
}
\examples{
f <- function(x){
x1 = x[1]
x2 = x[2]
ifelse(0<x1 & x1<1 & 0<x2 & x2<1 & 0<x1+x2 & x1+x2<1, 24*x1*x2, 0)}
a <- twoDsample(f , N=1000, lbx = 0, ubx = 1, lby = 0, uby = 1)
ggplot(a, aes(x, y)) +  geom_density_2d()

}
